+++ mrv2/lib/mrvGL/mrvGLUtil.h
@@ -59,6 +59,11 @@ namespace mrv
         render->drawMesh(mesh, pos, color);
     }
 
+    //! Draw a points in raster coordinates with glPointSize
+    void drawPoints(
+        const std::vector<math::Vector2f>& pts, const imaging::Color4f& color,
+        const int size = 1);
+
     //! Draw a single line in raster coordinates with a mesh.
     void drawLine(
         const std::shared_ptr<timeline::IRender>& render,