+++ mrv2/lib/mrvGL/mrvGLViewportPrivate.h
@@ -28,6 +28,7 @@ namespace mrv
         std::shared_ptr<tl::gl::OffscreenBuffer> annotation = nullptr;
         std::shared_ptr<tl::gl::Render> render = nullptr;
         std::shared_ptr<tl::gl::Shader> shader = nullptr;
+        std::shared_ptr<tl::gl::Shader> stereoShader = nullptr;
         std::shared_ptr<tl::gl::Shader> annotationShader = nullptr;
         int index = 0;
         int nextIndex = 1;